# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_EventView_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventView"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[EventView]",
        "type": "object"
      },
      "Page_EvidenceView_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceView"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[EvidenceView]",
        "type": "object"
      },
...
      "TaskSource": {
        "enum": [
          "Alert",
          "Ticket",
          "Schedule",
          "State",
          "Prediction",
          "Release",
          "Human",
          "AI"
        ],
        "title": "TaskSource",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "ops_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Weipai AI Ops Brain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/events": {
      "get": {
        "operationId": "list_events_api_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "prometheus",
                    "kubernetes",
                    "ops_platform",
                    "git",
                    "ci",
                    "argocd",
                    "config_center",
                    "cloud",
                    "manual",
                    "schedule",
                    "state",
                    "prediction",
                    "learning"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "query",
            "name": "service_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EventView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Events",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/events/{event_id}": {
      "get": {
        "operationId": "event_detail_api_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Event Detail",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/evidence": {
      "get": {
        "operationId": "list_evidence_api_evidence_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EvidenceView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Evidence",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/evidence/{evidence_id}": {
      "get": {
        "operationId": "evidence_detail_api_evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Evidence Detail",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/incidents": {
      "get": {
        "operationId": "list_incidents_api_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "service_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_IncidentHit_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Incidents",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/incidents/{incident_id}": {
      "get": {
        "operationId": "incident_detail_api_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentHit"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Incident Detail",
        "tags": [
          "任务闭环"
        ]
      }
    },
...
    "/api/status-history/{history_id}": {
      "get": {
        "operationId": "history_detail_api_status_history__history_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "history_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "History Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "History Detail",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/tasks": {
      "get": {
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TaskView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Tasks",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "operationId": "task_detail_api_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Task Detail",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/tasks/{task_id}/approval": {
      "post": {
        "operationId": "approve_api_tasks__task_id__approval_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/tasks/{task_id}/evidence": {
      "get": {
        "operationId": "task_evidence_api_tasks__task_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EvidenceView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Task Evidence",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/tasks/{task_id}/information": {
      "post": {
        "operationId": "inform_api_tasks__task_id__information_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Inform",
        "tags": [
          "任务闭环"
        ]
      }
    },
    "/api/tasks/{task_id}/interaction": {
      "get": {
        "operationId": "interaction_api_tasks__task_id__interaction_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Interaction",
        "tags": [
          "任务闭环"
        ]
      }
    },
...
  }
}